+++ client/lib/models/product.dart
@@ -18,15 +18,11 @@ class Product extends HiveObject {
   @HiveField(3)
   final String manufacturer;
 
-  @HiveField(4)
-  final String? photoUrl;
-
   Product({
     required this.barcode,
     required this.name,
     required this.unit,
     required this.manufacturer,
-    this.photoUrl,
   });
 
   factory Product.fromJson(Map<String, dynamic> json) =>